[docker-ci] Switch to relative paths for scp
When running on Mac OS I've encountered a protocol incompatiblity when packer
tries to scp files over to a freshly created instance. The symptom for this
issue is an error of the form

```
Retryable error: Error uploading script: scp: ./C:\button-down.ps1: Protocol
not available
```

To solve this issue I've switched to using relative paths for all files that
packer copies.